Update index.html
I added the head tags which as the tutorial said is a child of the html tags and a sibling to the body tags. this means that the head tags and the body tags must reside whithin an html opening and closing tag. I also added a title inside the head tags which will give the tab for this webpage a title as  well as this web page itself.